migrate /api/book/request-reschedule to viewer.booking.requestReschedule in trpc
